refactor: rework all imports and exports to ESM
