+++ lib/network/Network.js
@@ -9,7 +9,7 @@ import {
 } from "vis-util/esnext";
 import { DOTToGraph } from "./dotparser";
 import { parseGephi } from "./gephiParser";
-import Activator from "../shared/Activator";
+import { Activator } from "../shared/Activator";
 import * as locales from "./locales";
 import { normalizeLanguageCode } from "./locale-utils";
 
@@ -26,8 +26,8 @@ import InteractionHandler from "./modules/InteractionHandler";
 import SelectionHandler from "./modules/SelectionHandler";
 import LayoutEngine from "./modules/LayoutEngine";
 import ManipulationSystem from "./modules/ManipulationSystem";
-import Configurator from "./../shared/Configurator";
-import { Validator, printStyle } from "./../shared/Validator";
+import { Configurator } from "../shared/Configurator";
+import { VALIDATOR_PRINT_STYLE, Validator } from "../shared/Validator";
 import { allOptions, configureOptions } from "./options.js";
 import KamadaKawai from "./modules/KamadaKawai.js";
 
@@ -184,7 +184,7 @@ Network.prototype.setOptions = function (options) {
     if (errorFound === true) {
       console.error(
         "%cErrors have been found in the supplied options object.",
-        printStyle
+        VALIDATOR_PRINT_STYLE
       );
     }
