+++ lib/network/modules/components/Node.js
@@ -25,7 +25,7 @@ import Star from "./nodes/shapes/Star";
 import Text from "./nodes/shapes/Text";
 import Triangle from "./nodes/shapes/Triangle";
 import TriangleDown from "./nodes/shapes/TriangleDown";
-import { printStyle } from "../../../shared/Validator";
+import { VALIDATOR_PRINT_STYLE } from "../../../shared/Validator";
 
 /**
  * A node. A node can be connected to other nodes via one or multiple edges.
@@ -841,7 +841,7 @@ class Node {
       }
       console.error(
         "%cNegative or zero mass disallowed" + strId + ", setting mass to 1.",
-        printStyle
+        VALIDATOR_PRINT_STYLE
       );
       options.mass = 1;
     }
