+++ lib/shared/Activator.js
@@ -14,7 +14,7 @@ import Hammer from "../module/hammer";
  * @param {Element} container
  * @class Activator
  */
-function Activator(container) {
+export function Activator(container) {
   this._cleanupQueue = [];
 
   this.active = false;
@@ -166,6 +166,3 @@ function _hasParent(element, parent) {
   }
   return false;
 }
-
-export default Activator;
-export { Activator };
