+++ lib/shared/ColorPicker.js
@@ -158,7 +158,7 @@ const htmlColors = {
 /**
  * @param {number} [pixelRatio=1]
  */
-class ColorPicker {
+export class ColorPicker {
   /**
    * @param {number} [pixelRatio=1]
    */
@@ -802,5 +802,3 @@ class ColorPicker {
       ")";
   }
 }
-
-export default ColorPicker;
