+++ lib/shared/Configurator.js
@@ -2,7 +2,7 @@ import "./Configurator.css";
 
 import { copyAndExtendArray } from "vis-util/esnext";
 
-import ColorPicker from "./ColorPicker";
+import { ColorPicker } from "./ColorPicker";
 
 /**
  * Wrap given text (last argument) in HTML elements (all preceding arguments).
@@ -32,7 +32,7 @@ function wrapInTag(...rest) {
  *
  * The options are matched with their counterparts in each of the modules and the values used in the configuration are
  */
-class Configurator {
+export class Configurator {
   /**
    * @param {object} parentModule        | the location where parentModule.setOptions() can be called
    * @param {object} defaultContainer    | the default container of the module
@@ -809,5 +809,3 @@ class Configurator {
     return options;
   }
 }
-
-export default Configurator;
