+++ lib/shared/Popup.js
@@ -3,7 +3,7 @@ import "./Popup.css";
 /**
  * Popup is a class to create a popup window with some text
  */
-class Popup {
+export class Popup {
   /**
    * @param {Element} container       The container object.
    * @param {string}  overflowMethod  How the popup should act to overflowing ('flip' or 'cap')
@@ -134,5 +134,3 @@ class Popup {
     this.frame.parentNode.removeChild(this.frame); // Remove element from DOM
   }
 }
-
-export default Popup;
