+++ lib/shared/Validator.js
@@ -2,11 +2,13 @@ import { copyAndExtendArray, copyArray } from "vis-util/esnext";
 
 let errorFound = false;
 let allOptions;
-const printStyle = "background: #FFeeee; color: #dd0000";
+
+export const VALIDATOR_PRINT_STYLE = "background: #FFeeee; color: #dd0000";
+
 /**
  *  Used to validate options.
  */
-class Validator {
+export class Validator {
   /**
    * @ignore
    */
@@ -125,7 +127,7 @@ class Validator {
     const log = function (message) {
       console.error(
         "%c" + message + Validator.printLocation(path, option),
-        printStyle
+        VALIDATOR_PRINT_STYLE
       );
     };
 
@@ -270,7 +272,7 @@ class Validator {
 
     console.error(
       '%cUnknown option detected: "' + option + '"' + msg,
-      printStyle
+      VALIDATOR_PRINT_STYLE
     );
     errorFound = true;
   }
@@ -419,5 +421,3 @@ class Validator {
     return matrix[b.length][a.length];
   }
 }
-
-export { Validator, printStyle };
