+++ test/Label.test.js
@@ -16,7 +16,7 @@ import assert from "assert";
 import * as util from "vis-util/esnext";
 import { DataSet } from "vis-data/esnext";
 
-import canvasMockify from "./canvas-mock";
+import { canvasMockify } from "./canvas-mock";
 
 import Label from "../lib/network/modules/components/shared/Label";
 import NodesHandler from "../lib/network/modules/NodesHandler";
