+++ test/Network.test.js
@@ -14,8 +14,8 @@ import fs from "fs";
 import assert from "assert";
 import { DataSet } from "vis-data/esnext";
 import Network from "../lib/network/Network";
-import canvasMockify from "./canvas-mock";
-import { allOptions, configureOptions } from "./../lib/network/options.js";
+import { canvasMockify } from "./canvas-mock";
+import { allOptions, configureOptions } from "../lib/network/options.js";
 
 /**
  * Merge all options of object b into object b
